refactor(blog): remove back.tsx and integrate back button directly in page.tsx
feat(blog): create a dedicated not-found component for reuse across pages

refactor(blog): update component routes to use new structure

refactor(command-menu): update href for component posts to new route structure

refactor(not-found): centralize not-found page logic in a single component for consistency and reuse
